[refactor] change metrics dir structure
