--- opencda/core/plan/metrics/__init__.py
@@ -1,8 +1,0 @@
-from opencda.core.plan.metrics.base_metric import BaseMetric
-from opencda.core.plan.metrics.dynamics_metric import DynamicsMetric
-from opencda.core.plan.metrics.metric_collector import MetricCollector
-from opencda.core.plan.metrics.metric_sample import MetricSample
-from opencda.core.plan.metrics.registry import MetricRegistry
-from opencda.core.plan.metrics.ttc_metric import TtcMetric
-
-__all__ = ["BaseMetric", "DynamicsMetric", "MetricCollector", "MetricRegistry", "MetricSample", "TtcMetric"]
